% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconvs.R
\name{saf_buac_to_kgha_corn}
\alias{saf_buac_to_kgha_corn}
\title{saf_lbac_to_kgha}
\usage{
saf_buac_to_kgha_corn(dat_buac)
}
\arguments{
\item{dat_buac}{A vector of corn yields in bushels per acre}
}
\value{
A vector of corn yields in kilograms per hectare (adjusted for the 15% moisture the bu/ac are reported at)
}
\description{
saf_lbac_to_kgha
}
