% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizNrespcurves.R
\name{saf_combN}
\alias{saf_combN}
\title{Combine experimental and predicted data}
\usage{
saf_combN(exp_data, pred_data)
}
\arguments{
\item{exp_data}{Experimental data, see sad_exdata for format example, yield must be in kg/ha}

\item{pred_data}{Predicted data, output from saf_predNresp, yield must be in kg/ha}
}
\value{
A tibble with columns named data_type (pred or exp) and yield_kgha
}
\description{
Combine experimental and predicted data
}
