% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateconvs.R
\name{saf_date_to_doy}
\alias{saf_date_to_doy}
\title{saf_date_to_doy}
\usage{
saf_date_to_doy(mydate = "2001-01-01")
}
\arguments{
\item{mydate}{A date in yyyy-mm-dd format}
}
\value{
A numeric value representing the day of year, doy
}
\description{
saf_date_to_doy
}
