% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateconvs.R
\name{saf_doy_to_date}
\alias{saf_doy_to_date}
\title{saf_doy_to_date}
\usage{
saf_doy_to_date(mydoy = 1, myyear = 2000)
}
\arguments{
\item{mydoy}{A day of year}

\item{myyear}{A year if known, defaults to 2000}
}
\value{
A date in the format yyyy-mm-dd
}
\description{
saf_doy_to_date
}
