% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNresp.R
\name{saf_fitNresp}
\alias{saf_fitNresp}
\title{saf_fitNresp}
\usage{
saf_fitNresp(mydata)
}
\arguments{
\item{mydata}{The N resp data that has been grouped (site, year, rotation) and nested. See the included data 'sad_exdata' for an example}
}
\value{
A tibble with the data and fits nested
}
\description{
saf_fitNresp
}
