% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconvs.R
\name{saf_kgha_to_buac_corn}
\alias{saf_kgha_to_buac_corn}
\title{saf_kgha_to_buac_corn}
\usage{
saf_kgha_to_buac_corn(dat_kgha)
}
\arguments{
\item{dat_kgha}{A vector of corn yields in bushels per acre}
}
\value{
A vector of corn yields in bushels per acre; adjusted for the 15\% moisture the bu/ac are reported at
}
\description{
saf_kgha_to_buac_corn
}
