% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconvs.R
\name{saf_lbac_to_kgha}
\alias{saf_lbac_to_kgha}
\title{saf_lbac_to_kgha}
\usage{
saf_lbac_to_kgha(dat_lbac)
}
\arguments{
\item{dat_lbac}{A vector of values in pounds per acre}
}
\value{
A vector of values in kilograms per hectare (multiplied by 1.123)
}
\description{
saf_lbac_to_kgha
}
