% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predNresp.R
\name{saf_predNresp}
\alias{saf_predNresp}
\title{saf_predNresp}
\usage{
saf_predNresp(mydata, highN = 270)
}
\arguments{
\item{mydata}{Results from saf_fitNresp() function}

\item{highN}{Highest rate of nitrogen application (in kg/ha) you want to predict (default is 270)}
}
\value{
A tibble with the predictions for 0-highN by 1 kg/ha
}
\description{
saf_predNresp
}
