% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readapout.R
\name{saf_readapout}
\alias{saf_readapout}
\title{saf_readapout}
\usage{
saf_readapout(fold_dir)
}
\arguments{
\item{fold_dir}{A character string of the file path to the folder with the APSIM out files in it}
}
\value{
A tibble with the data from all the out files (you will need to filter them)
}
\description{
saf_readapout
}
\examples{
No examples yet
}
