% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNrespcurves.R
\name{saf_smyNfits}
\alias{saf_smyNfits}
\title{saf_smyNfits}
\usage{
saf_smyNfits(myfits)
}
\arguments{
\item{myfits}{The resulting tibble from using the saf_fitNresp function}
}
\value{
A tibble (nothing nested) with the fitted parameters for each model, plus the R2 value
}
\description{
saf_smyNfits
}
\examples{
No examples yet
}
