% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizNrespcurves.R
\name{saf_vizNexppred}
\alias{saf_vizNexppred}
\title{Quick viz of experimental data points (field or apsim) and statistical model preds}
\usage{
saf_vizNexppred(comb_data)
}
\arguments{
\item{comb_data}{Output from saf_combN}
}
\value{
A plot faceted by model type and site
}
\description{
Quick viz of experimental data points (field or apsim) and statistical model preds
}
